Summary service + exception handling and refactor

+++ src/main/java/com/nmkip/weather/repository/SummaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SummaryRepository extends CrudRepository<Summary, Integer> {
+public interface SummaryRepository extends CrudRepository<Summary, Long> {
 }
